@@ -1,6 +1,8 @@
+import './env.mjs'
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
+  //
 }
 
-module.exports = nextConfig
+export default nextConfig
